requirements.txt
python-dotenv
ollama
